requirements.txt
Flask==3.0.3
gunicorn==22.0.0
flask-cors==4.0.1
gspread==6.1.2
google-auth==2.35.0
